+++ uninstall.php
@@ -19,6 +19,8 @@ delete_option( 'mmsar_content_signals' );
 delete_option( 'mmsar_structured_data' );
 delete_option( 'mmsar_features' );
 delete_option( 'mmsar_endpoints' );
+delete_option( 'mmsar_agent_log' );
+delete_option( 'mmsar_agent_log_pages' );
 delete_transient( 'llmmd_llms_txt' );
 delete_transient( 'mmsar_llms_full_txt' );
 delete_transient( 'mmsar_flush_needed' );